requirements.txt
altgraph==0.17.5
ImageIO==2.37.2
lazy_loader==0.4
networkx==3.6.1
numpy==2.4.1
opencv-python==4.13.0.90
packaging==26.0
pillow==12.1.0
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
scikit-image==0.26.0
scipy==1.17.0
setuptools==80.10.2
shiboken6==6.10.1
tifffile==2026.1.28
